modifications about article upload images.webuploader cannot work on some browers.

+++ zcms/src/com/mavict/article/ArticleDao.java
@@ -22,7 +22,7 @@ public interface ArticleDao extends BaseDao<Article, Long> {
 	 * 
 	 * @return 文章实体List
 	 * */
-	List<Article> getNumList(Integer categoryId,Integer Num);
+	List<Article> getNumList(Long categoryId,Integer Num);
 	
 	/**
 	 * 获取指定目录下的文章分页操作
@@ -32,12 +32,12 @@ public interface ArticleDao extends BaseDao<Article, Long> {
 	 * 
 	 * @return 
 	 * */
-	PagedContent<Article> getPagedContentByCategoryId(Integer categoryId,PageInfo pageInfo);
+	PagedContent<Article> getPagedContentByCategoryId(Long categoryId,PageInfo pageInfo);
 	
 	/**
 	 * 获取指定目录下文章的数目
 	 * 
 	 * @param categoryId 文章目录Id
 	 * */
-	Long count(Integer categoryId);
+	Long count(Long categoryId);
 }
